+++ test/PHPUnit/pastell-core/type-dossier/TypeDossierEtapeDefinitionTest.php
@@ -165,6 +165,7 @@ class TypeDossierEtapeDefinitionTest extends PastellTestCase
             'annotation_privee' => 'annotation_privee_2',
             'primo_signature_detachee' => 'primo_signature_detachee_2',
             'parapheur_date_signature' => 'parapheur_date_signature_2',
+            'iparapheur_metadata_sortie' => 'iparapheur_metadata_sortie_2',
         ], $mapping);
     }
 
@@ -220,6 +221,7 @@ class TypeDossierEtapeDefinitionTest extends PastellTestCase
                             'name' => 'Métadonnées parapheur (JSON)',
                             'commentaire' => 'Au format JSON {"cle1":"valeur1","cle2":"valeur2",...}',
                             'type' => 'file',
+                            'visionneuse' => 'Pastell\Viewer\JsonViewer',
                         ],
                     'has_date_limite_2' =>
                         [
@@ -307,6 +309,13 @@ class TypeDossierEtapeDefinitionTest extends PastellTestCase
                             'name' => 'Date de dernière signature',
                             'type' => 'date',
                             'read-only' => true,
+                        ],
+                    'iparapheur_metadata_sortie_2' =>
+                        [
+                            'name' => 'Métadonnées de sortie du parapheur',
+                            'type' => 'file',
+                            'read-only' => true,
+                            'visionneuse' => 'Pastell\Viewer\JsonViewer',
                         ],
                 ],
             'Parapheur FAST #2' => [
@@ -418,7 +427,8 @@ Uniquement avec le mode "circuit à la volée"',
                         'rejet-iparapheur' => 'rejet-iparapheur_2',
                         'erreur-verif-iparapheur' => 'erreur-verif-iparapheur_2',
                         'parapheur_last_message' => 'parapheur_last_message_2',
-                        'parapheur_date_signature' => 'parapheur_date_signature_2'
+                        'parapheur_date_signature' => 'parapheur_date_signature_2',
+                        'iparapheur_metadata_sortie' => 'iparapheur_metadata_sortie_2',
                     ],
                 ],
                 'erreur-verif-iparapheur_2' => [